Remove TriangleScale and Implement TriangleVelocity
The main reason for this was that scaling would not take effect unless you restarted the client completely, and had the config and its value changed. This is inconvenient, and as such, has been temporarily removed until we can sort of out how refresh its state onChange. TriangleSpeed simply allows velocity shifting all across, and inherently works with everything onChange.